requirements.txt
tensorflow-cpu==2.16.1
Flask==3.0.1
numpy==1.25.2
Pillow==10.0.1
Werkzeug==3.0.1
# Optional/utility
gunicorn==20.1.0
